Fix expiration_time JSON type mapping
Change expiration_time from int64 to int32 in protobuf schema so that protojson serializes it as an integer instead of a string, preserving exact JSON schema compatibility with existing API consumers.

+++ keymanager/workload_service/server.go
@@ -393,7 +393,7 @@ func (s *Server) generateKEMKey(w http.ResponseWriter, req *api.GenerateKeyReque
 			PublicKey: base64.StdEncoding.EncodeToString(kemPubKey),
 		},
 		KeyProtectionMechanism: keymanager.KeyProtectionMechanism_KEY_PROTECTION_VM_EMULATED.String(),
-		ExpirationTime:         time.Now().Unix() + int64(req.Lifespan),
+		ExpirationTime:         int32(time.Now().Unix()) + int32(req.Lifespan),
 	}
 	writeJSON(w, &resp, http.StatusOK)
 }
@@ -448,7 +448,7 @@ func (s *Server) handleEnumerateKeys(w http.ResponseWriter, _ *http.Request) {
 				PublicKey: base64.StdEncoding.EncodeToString(key.KEMPubKey),
 			},
 			KeyProtectionMechanism: keymanager.KeyProtectionMechanism_KEY_PROTECTION_VM_EMULATED.String(),
-			ExpirationTime:         time.Now().Unix() + int64(key.RemainingLifespanSecs),
+			ExpirationTime:         int32(time.Now().Unix()) + int32(key.RemainingLifespanSecs),
 		})
 	}
